requirements.txt
torch==2.5.1
torchvision==0.20.1
numpy==1.26.4
scipy==1.13.1
Pillow==11.0.0
tqdm==4.67.1
imageio==2.33.1
scikit-learn==1.1.3
matplotlib==3.8.4
streamlit==1.41.1
sklearn-pandas==2.2.0
spacy==3.4.4
html5lib==1.1
Jinja2==3.1.3
